LPD-17341 DDLRecordBatchReindexer is an internal interface, it should not be registered as osgi service
