+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/internal/search/DDLRecordSetModelSearchConfigurator.java
@@ -6,10 +6,13 @@
 package com.liferay.dynamic.data.lists.internal.search;
 
 import com.liferay.dynamic.data.lists.internal.search.spi.model.index.contributor.DDLRecordSetModelIndexerWriterContributor;
+import com.liferay.dynamic.data.lists.model.DDLRecord;
 import com.liferay.dynamic.data.lists.model.DDLRecordSet;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetLocalService;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.search.batch.DynamicQueryBatchIndexingActionableFactory;
+import com.liferay.portal.search.indexer.IndexerDocumentBuilder;
+import com.liferay.portal.search.indexer.IndexerWriter;
 import com.liferay.portal.search.spi.model.index.contributor.ModelIndexerWriterContributor;
 import com.liferay.portal.search.spi.model.registrar.ModelSearchConfigurator;
 
@@ -53,12 +56,11 @@ public class DDLRecordSetModelSearchConfigurator
 	protected void activate() {
 		_modelIndexWriterContributor =
 			new DDLRecordSetModelIndexerWriterContributor(
-				_ddlRecordBatchReindexer, _ddlRecordSetLocalService,
+				new DDLRecordBatchReindexer(
+					_indexerDocumentBuilder, _indexerWriter),
+				_ddlRecordSetLocalService,
 				_dynamicQueryBatchIndexingActionableFactory);
 	}
-
-	@Reference
-	private DDLRecordBatchReindexer _ddlRecordBatchReindexer;
 
 	@Reference
 	private DDLRecordSetLocalService _ddlRecordSetLocalService;
@@ -66,6 +68,16 @@ public class DDLRecordSetModelSearchConfigurator
 	@Reference
 	private DynamicQueryBatchIndexingActionableFactory
 		_dynamicQueryBatchIndexingActionableFactory;
+
+	@Reference(
+		target = "(indexer.class.name=com.liferay.dynamic.data.lists.model.DDLRecord)"
+	)
+	private IndexerDocumentBuilder _indexerDocumentBuilder;
+
+	@Reference(
+		target = "(indexer.class.name=com.liferay.dynamic.data.lists.model.DDLRecord)"
+	)
+	private IndexerWriter<DDLRecord> _indexerWriter;
 
 	private ModelIndexerWriterContributor<DDLRecordSet>
 		_modelIndexWriterContributor;
